upate vit variant base
