+++ medicai/models/vit/__init__.py
@@ -1,2 +1,2 @@
-from .vision_transformer import ViT
+from .vit import ViT
 from .vit_backbone import ViTBackbone
